Organize components into feature folders
