+++ by-tomorrow/src/lib/components/editor/deleteAlert.svelte
